Additional changes to ITypeScannerConfigurator and TypeScannerConfigurator
