+++ src/MefContrib.Hosting.Conventions.Tests/TypeScannerConfiguratorTests.cs
@@ -12,36 +12,36 @@
     public class TypeScannerConfiguratorTests
     {
         [Test]
-        public void Add_should_throw_argumentnullexception_when_called_with_null()
+        public void Scanner_should_throw_argumentnullexception_when_called_with_null()
         {
             var configurator =
                 new TypeScannerConfigurator();
 
             var exception =
-                Catch.Exception(() => configurator.Add(null));
+                Catch.Exception(() => configurator.Scanner(null));
 
             exception.ShouldBeOfType<ArgumentNullException>();
         }
 
         [Test]
-        public void Add_should_return_reference_to_same_configurator()
+        public void Scanner_should_return_reference_to_same_configurator()
         {
             var configurator =
                 new TypeScannerConfigurator();
 
             var reference =
-                configurator.Add(new TypeScanner());
+                configurator.Scanner(new TypeScanner());
 
             reference.ShouldBeSameAs(configurator);
         }
 
         [Test]
-        public void Add_should_add_scanner_to_configurator()
+        public void Scanner_should_add_scanner_to_configurator()
         {
             var configurator =
                 new TypeScannerConfigurator();
 
-            configurator.Add(new TypeScanner(new[] { typeof(object) }));
+            configurator.Scanner(new TypeScanner(new[] { typeof(object) }));
 
             var scanner =
                 configurator.GetTypeScanner();
